add abstract to simple components
