+++ src/view/abstract.js
@@ -1,0 +1,27 @@
+import {createElement} from "../utils";
+
+export default class Abstract {
+  constructor() {
+    if (new.target === Abstract) {
+      throw new Error(`Can't instantiate Abstract, only concrete one.`);
+    }
+
+    this._element = null;
+  }
+
+  _getTemplate() {
+    throw new Error(`Abstract method not implemented: getTemplate`);
+  }
+
+  get element() {
+    if (!this._element) {
+      this._element = createElement(this._getTemplate());
+    }
+
+    return this._element;
+  }
+
+  removeElement() {
+    this._element = null;
+  }
+}
