+++ src/view/films-container.js
@@ -1,25 +1,10 @@
-import {createElement} from "../utils";
+import AbstractView from "./abstract";
 
-export default class FilmsContainer {
-  constructor() {
-    this._element = null;
-  }
+export default class FilmsContainer extends AbstractView {
 
   _getTemplate() {
     return (
       `<section class="films"></section>`
     );
-  }
-
-  get element() {
-    if (!this._element) {
-      this._element = createElement(this._getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
