+++ src/view/films-list.js
@@ -1,9 +1,6 @@
-import {createElement} from "../utils";
+import AbstractView from "./abstract";
 
-export default class FilmsList {
-  constructor() {
-    this._element = null;
-  }
+export default class FilmsList extends AbstractView {
 
   _getTemplate() {
     return (
@@ -15,15 +12,4 @@ export default class FilmsList {
     );
   }
 
-  get element() {
-    if (!this._element) {
-      this._element = createElement(this._getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
-  }
 }
