+++ src/view/menu-container.js
@@ -1,9 +1,6 @@
-import {createElement} from "../utils";
+import AbstractView from "./abstract";
 
-export default class MenuContainer {
-  constructor() {
-    this._element = null;
-  }
+export default class MenuContainer extends AbstractView {
 
   _getTemplate() {
     return (
@@ -11,14 +8,4 @@ export default class MenuContainer {
     );
   }
 
-  get element() {
-    if (!this._element) {
-      this._element = createElement(this._getTemplate());
-    }
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
-  }
 }
