+++ src/view/menu-statistic.js
@@ -1,25 +1,9 @@
-import {createElement} from "../utils";
+import AbstractView from "./abstract";
 
-export default class MainMenuStatistic {
-  constructor() {
-    this._element = null;
-  }
-
+export default class MainMenuStatistic extends AbstractView {
   _getTemplate() {
     return (
       `<a href="#stats" class="main-navigation__additional">Stats</a>`
     );
-  }
-
-  get element() {
-    if (!this._element) {
-      this._element = createElement(this._getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
