+++ src/view/no_film.js
@@ -1,23 +1,8 @@
-import {createElement} from "../utils";
+import AbstractView from "./abstract";
 
-export default class NoFilm {
-  constructor() {
-    this._element = null;
-  }
+export default class NoFilm extends AbstractView {
 
   _getTemplate() {
     return `<h2 class="films-list__title">There are no movies in our database</h2>`;
-  }
-
-  get element() {
-    if (!this._element) {
-      this._element = createElement(this._getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
